Moved job classes to a models folder
