+++ batch/models/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+var JobSimple = require('./job_simple');
+var JobMultiple = require('./job_multiple');
+
+module.exports = [ JobSimple, JobMultiple ];
